[feat] product_detail 데이터 받아오기 기능 완료
