+++ client/css/product-detail.css
@@ -254,7 +254,15 @@
 .product-menu {
   display: flex;
   margin-bottom: 40px;
+  z-index: 10;
 }
+.drop{
+  position: fixed;
+  margin: 0 auto;
+  top: 0;
+  width: 100%;
+}
+
 
 .product-menu_nav {
   width: 263.25px;
@@ -587,7 +595,8 @@
 .product-review_arrow-left {
   width: 56px;
   height: 56px;
-  background: url("/client/assets/icons/web-icons/Arrow-2.svg") no-repeat;
+  /* background: url("/client/assets/icons/web-icons/Arrow-2.svg") no-repeat; */
+  background: url(../assets/icons/web-icons/Arrow-2.svg) no-repeat;
   background-position: -20px -165px;
   background-size: 95px;
   border: 1px solid var(--gray-100);
@@ -597,7 +606,7 @@
 .product-review_arrow-right {
   width: 56px;
   height: 56px;
-  background: url("/client/assets/icons/web-icons/Arrow-2.svg") no-repeat;
+  background: url(../assets/icons/web-icons/Arrow-2.svg) no-repeat;
   background-position: -20px -238px;
   background-size: 95px;
   border: 1px solid var(--gray-100);
@@ -630,7 +639,7 @@
   padding: 32px;
   position: relative;
   width: 793px;
-  height: 630px;
+  height: 550px;
   border-radius: 16px;
   overflow:auto;
   background: var(--white)
@@ -796,14 +805,14 @@
   height: 19px;
   position: absolute;
   right: 40px;
-  bottom: 150px;
+  bottom: 73px;
   font-weight: 400;
   font-size: 12.003px;
   color: var(--gray-300);
 }
 .forminput-textarea_limit-2 {
   right: 35px;
-  bottom: 150px;
+  bottom: 70px;
 }
 .forminput-buttons {
   margin-top: 20px;
